requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
albumentations>=1.3.0
grad-cam>=1.4.8
pandas>=2.0.0
gradio>=4.0.0
huggingface-hub>=0.16.0
umap-learn>=0.5.3
opencv-python-headless>=4.8.0
